refactor(query): migrate profile to common base
