+++ src/common/base/src/lib.rs
@@ -24,6 +24,7 @@
 #![feature(alloc_error_hook)]
 #![feature(slice_swap_unchecked)]
 #![feature(lint_reasons)]
+#![feature(variant_count)]
 
 pub mod base;
 pub mod containers;
