+++ src/common/base/src/runtime/profile/mod.rs
@@ -1,0 +1,23 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#[allow(clippy::module_inception)]
+mod profile;
+mod profiles;
+
+pub use profile::Profile;
+pub use profile::ProfileLabel;
+pub use profiles::get_statistics_desc;
+pub use profiles::ProfileDesc;
+pub use profiles::ProfileStatisticsName;
