+++ src/query/catalog/src/table_context.rs
@@ -24,6 +24,7 @@ use std::time::SystemTime;
 use dashmap::DashMap;
 use databend_common_base::base::Progress;
 use databend_common_base::base::ProgressValues;
+use databend_common_base::runtime::profile::Profile;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_expression::DataBlock;
@@ -36,7 +37,6 @@ use databend_common_meta_app::principal::RoleInfo;
 use databend_common_meta_app::principal::UserDefinedConnection;
 use databend_common_meta_app::principal::UserInfo;
 use databend_common_pipeline_core::processors::PlanProfile;
-use databend_common_pipeline_core::processors::Profile;
 use databend_common_pipeline_core::InputError;
 use databend_common_settings::Settings;
 use databend_common_storage::CopyStatus;
