+++ src/query/pipeline/core/Cargo.toml
@@ -20,7 +20,6 @@ async-backtrace = { workspace = true }
 async-trait = { workspace = true }
 futures = { workspace = true }
 minitrace = { workspace = true }
-once_cell = { workspace = true }
 petgraph = "0.6.2"
 serde = { workspace = true }
