+++ src/query/pipeline/core/src/processors/mod.rs
@@ -16,17 +16,12 @@ mod port;
 mod processor;
 
 mod duplicate_processor;
-mod plan_profile;
 mod port_trigger;
-pub mod profile;
-mod profiles;
+mod profile;
 mod resize_processor;
 mod shuffle_processor;
 
 pub use duplicate_processor::DuplicateProcessor;
-pub use plan_profile::PlanProfile;
-pub use plan_profile::PlanScope;
-pub use plan_profile::PlanScopeGuard;
 pub use port::connect;
 pub use port::InputPort;
 pub use port::OutputPort;
@@ -37,11 +32,9 @@ pub use processor::Event;
 pub use processor::EventCause;
 pub use processor::Processor;
 pub use processor::ProcessorPtr;
-pub use profile::Profile;
-pub use profile::ProfileLabel;
-pub use profiles::get_statistics_desc;
-pub use profiles::ProfileDesc;
-pub use profiles::ProfileStatisticsName;
+pub use profile::PlanProfile;
+pub use profile::PlanScope;
+pub use profile::PlanScopeGuard;
 pub use resize_processor::create_resize_item;
 pub use resize_processor::ResizeProcessor;
 pub use shuffle_processor::ShuffleProcessor;
