+++ src/query/pipeline/sources/src/input_formats/source_aligner.rs
@@ -19,6 +19,8 @@ use std::sync::Arc;
 
 use databend_common_base::base::tokio::sync::mpsc::Receiver;
 use databend_common_base::base::ProgressValues;
+use databend_common_base::runtime::profile::Profile;
+use databend_common_base::runtime::profile::ProfileStatisticsName;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_expression::DataBlock;
@@ -26,8 +28,6 @@ use databend_common_pipeline_core::processors::Event;
 use databend_common_pipeline_core::processors::OutputPort;
 use databend_common_pipeline_core::processors::Processor;
 use databend_common_pipeline_core::processors::ProcessorPtr;
-use databend_common_pipeline_core::processors::Profile;
-use databend_common_pipeline_core::processors::ProfileStatisticsName;
 use log::debug;
 
 use crate::input_formats::input_pipeline::AligningStateTrait;
