+++ src/query/service/src/api/rpc/exchange/exchange_manager.rs
@@ -22,6 +22,7 @@ use async_channel::Receiver;
 use databend_common_arrow::arrow_format::flight::data::FlightData;
 use databend_common_arrow::arrow_format::flight::service::flight_service_client::FlightServiceClient;
 use databend_common_base::base::GlobalInstance;
+use databend_common_base::runtime::profile::Profile;
 use databend_common_base::runtime::GlobalIORuntime;
 use databend_common_base::runtime::Thread;
 use databend_common_base::runtime::TrySpawn;
@@ -30,7 +31,6 @@ use databend_common_config::GlobalConfig;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_grpc::ConnectionFactory;
-use databend_common_pipeline_core::processors::profile::Profile;
 use databend_common_sql::executor::PhysicalPlan;
 use minitrace::prelude::*;
 use parking_lot::Mutex;
