+++ src/query/service/src/api/rpc/exchange/statistics_sender.rs
@@ -20,11 +20,11 @@ use std::time::Duration;
 use async_channel::Sender;
 use databend_common_base::base::tokio::task::JoinHandle;
 use databend_common_base::base::tokio::time::sleep;
+use databend_common_base::runtime::profile::Profile;
 use databend_common_base::runtime::TrySpawn;
 use databend_common_catalog::table_context::TableContext;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
-use databend_common_pipeline_core::processors::profile::Profile;
 use databend_common_pipeline_core::processors::PlanProfile;
 use databend_common_storage::MergeStatus;
 use futures_util::future::Either;
