+++ src/query/service/src/interpreters/interpreter.rs
@@ -16,14 +16,14 @@ use std::collections::BTreeMap;
 use std::sync::Arc;
 use std::time::SystemTime;
 
+use databend_common_base::runtime::profile::get_statistics_desc;
+use databend_common_base::runtime::profile::ProfileDesc;
+use databend_common_base::runtime::profile::ProfileStatisticsName;
 use databend_common_catalog::table_context::TableContext;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_expression::SendableDataBlockStream;
-use databend_common_pipeline_core::get_statistics_desc;
 use databend_common_pipeline_core::processors::PlanProfile;
-use databend_common_pipeline_core::processors::ProfileDesc;
-use databend_common_pipeline_core::processors::ProfileStatisticsName;
 use databend_common_pipeline_core::SourcePipeBuilder;
 use log::error;
 use log::info;
