+++ src/query/service/src/pipelines/executor/processor_async_task.rs
@@ -24,11 +24,11 @@ use std::time::Instant;
 
 use databend_common_base::base::tokio::time::sleep;
 use databend_common_base::runtime::catch_unwind;
+use databend_common_base::runtime::profile::Profile;
+use databend_common_base::runtime::profile::ProfileStatisticsName;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_pipeline_core::processors::ProcessorPtr;
-use databend_common_pipeline_core::processors::Profile;
-use databend_common_pipeline_core::processors::ProfileStatisticsName;
 use futures_util::future::BoxFuture;
 use futures_util::future::Either;
 use futures_util::FutureExt;
