+++ src/query/service/src/pipelines/processors/transforms/transform_sort_spill.rs
@@ -18,6 +18,8 @@ use std::marker::PhantomData;
 use std::sync::Arc;
 use std::time::Instant;
 
+use databend_common_base::runtime::profile::Profile;
+use databend_common_base::runtime::profile::ProfileStatisticsName;
 use databend_common_exception::Result;
 use databend_common_expression::types::DataType;
 use databend_common_expression::types::NumberDataType;
@@ -38,8 +40,6 @@ use databend_common_pipeline_core::processors::Event;
 use databend_common_pipeline_core::processors::InputPort;
 use databend_common_pipeline_core::processors::OutputPort;
 use databend_common_pipeline_core::processors::Processor;
-use databend_common_pipeline_core::processors::Profile;
-use databend_common_pipeline_core::processors::ProfileStatisticsName;
 use databend_common_pipeline_transforms::processors::sort::CommonRows;
 use databend_common_pipeline_transforms::processors::sort::DateRows;
 use databend_common_pipeline_transforms::processors::sort::HeapMerger;
