+++ src/query/service/src/sessions/session_mgr.rs
@@ -24,13 +24,13 @@ use std::time::Duration;
 use databend_common_base::base::tokio;
 use databend_common_base::base::GlobalInstance;
 use databend_common_base::base::SignalStream;
+use databend_common_base::runtime::profile::Profile;
 use databend_common_catalog::table_context::ProcessInfoState;
 use databend_common_config::GlobalConfig;
 use databend_common_config::InnerConfig;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_metrics::session::*;
-use databend_common_pipeline_core::processors::profile::Profile;
 use databend_common_settings::Settings;
 use futures::future::Either;
 use futures::StreamExt;
