+++ src/query/storages/parquet/src/parquet2/processors/deserialize_transform.rs
@@ -22,6 +22,8 @@ use databend_common_arrow::arrow::bitmap::MutableBitmap;
 use databend_common_arrow::parquet::indexes::Interval;
 use databend_common_base::base::Progress;
 use databend_common_base::base::ProgressValues;
+use databend_common_base::runtime::profile::Profile;
+use databend_common_base::runtime::profile::ProfileStatisticsName;
 use databend_common_catalog::plan::PartInfoPtr;
 use databend_common_catalog::query_kind::QueryKind;
 use databend_common_catalog::table_context::TableContext;
@@ -42,8 +44,6 @@ use databend_common_pipeline_core::processors::InputPort;
 use databend_common_pipeline_core::processors::OutputPort;
 use databend_common_pipeline_core::processors::Processor;
 use databend_common_pipeline_core::processors::ProcessorPtr;
-use databend_common_pipeline_core::processors::Profile;
-use databend_common_pipeline_core::processors::ProfileStatisticsName;
 use databend_common_storage::CopyStatus;
 use databend_common_storage::FileStatus;
 use opendal::services::Memory;
